Revert "Define a default background color"
Defining a background color seemed to solve the problem with the search
input being effectively invisible, but it actually didn't.

It's as if maybe the search input takes on the default background so if
the current background is the same as whatever the default happens to
be, then we have that problem.

This reverts commit 9df3861b0feae74cd237d5e1eb6d4362527b0876.